+++ doc/changelog.rst
@@ -22,6 +22,7 @@ Bug fixes
 - Fix order printouts template by checking whether the addresses are valid
   before calling methods.
 - Fix front manufacturer modified filter to consider only visible shop products
+- Make Xtheme TextPlugin fallback to empty string when there is no translation for the current language.
 
 Front
 ~~~~~
@@ -44,6 +45,7 @@ Simple CMS
 Xtheme
 ~~~~~~
 
+- Make the Parler default language be the first language in plugin multi-language form
 - Add option to render extra templates in theme configuration.
 - Add option for per object placeholders
 - Fix bug deleting last row from placeholder